requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
psycopg2-binary
requests
pip